requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
boto==2.49.0
boto3==1.9.38
botocore==1.12.38
bz2file==0.98
certifi==2018.10.15
chainer==5.0.0
chardet==3.0.4
cycler==0.10.0
docutils==0.14
filelock==3.0.10
gensim==3.6.0
idna==2.7
isort==4.3.4
Janome==0.3.6
jmespath==0.9.3
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.1
mccabe==0.6.1
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
requests==2.20.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
smart-open==1.7.1
tqdm==4.28.1
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
